% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gmt.R
\name{format_gmt}
\alias{format_gmt}
\title{Format a gmt file into a list}
\usage{
format_gmt(path_to_gmt)
}
\arguments{
\item{path_to_gmt}{A path to the gmt file.}
}
\value{
A list where name is the term of the gmt and content is the ids
}
\description{
Take a path to a gmt file and format it into a easily readble/printing format
}
\examples{
path_to_gmt <- system.file("extdata", "h.all.v6.2.entrez.gmt",
                           package = "gprofiler2.addon", mustWork = TRUE)
formated_gmt <- format_gmt(path_to_gmt)

}
\author{
Gwenaëlle Lemoine <lemoine.gwenaelle@gmail.com>
}
