% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gost_aggreg_res.R
\name{gost_aggreg_res}
\alias{gost_aggreg_res}
\title{Aggregation of results from gost_standard and gost_custom_gmt}
\usage{
gost_aggreg_res(standard_output = NULL, custom_output = NULL,
  check_content = TRUE)
}
\arguments{
\item{standard_output}{output from \code{\link[gprofiler2]{gost}}}

\item{custom_output}{output from \code{\link{gost_custom_gmt}}}

\item{check_content}{boolean specifying if content must be checked to be from same queries}
}
\value{
A named list where 'result' contains data.frame with the enrichment analysis results and 'meta' contains metadata needed for Manhattan plot. If the input
 consisted of several lists the corresponding list is indicated with a variable
 'query'.
 When requesting a 'multi_query', either TRUE or FALSE, the columns of the resulting data frame differ.
 If 'evcodes' is set, the return value includes columns 'evidence_codes' and 'intersection'.
 The latter conveys info about the intersecting genes between the corresponding query and term.
}
\description{
Merging of the output from both \code{\link[gprofiler2]{gost}} and \code{\link{gost_custom_gmt}} on the same query or list of queries
May also be used on an output from this function with another output from \code{\link{gost_custom_gmt}}
}
\examples{
path_to_gmt <- system.file("extdata", "h.all.v6.2.entrez.gmt",
                           package = "gprofiler2.addon", mustWork = TRUE)
query <- c("26118", "5837", "6781", "23036", "694", "123", "1466", "7436",
           "23210", "2131", "2152", "5165", "55139", "7360", "229", "8614",
           "54206", "2027", "10957", "3162", "5228", "26330", "9435", "55076")
gost_custom_gmt_res <- gost_custom_gmt(query, custom_gmt = path_to_gmt)
gost_classic_res <- gprofiler2::gost(query, numeric_ns = "ENTREZGENE_ACC")
gost_aggregated <- gost_aggreg_res(gost_classic_res, gost_custom_gmt_res)

}
\author{
Gwenaëlle Lemoine <lemoine.gwenaelle@gmail.com>
}
