% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gost_custom_gmt.R
\name{gost_custom_gmt}
\alias{gost_custom_gmt}
\title{Gene list functional enrichment with custom gmt file}
\usage{
gost_custom_gmt(query, ordered_query = FALSE, multi_query = FALSE,
  significant = TRUE, exclude_iea = TRUE,
  measure_underrepresentation = FALSE, evcodes = FALSE,
  user_threshold = 0.05, correction_method = c("g_SCS", "bonferroni",
  "fdr", "false_discovery_rate", "gSCS", "analytical"),
  domain_scope = c("annotated", "known", "custom"), custom_bg = NULL,
  numeric_ns = "", custom_gmt = NULL)
}
\arguments{
\item{query}{vector that can consist of IDs of the same type as those inside gmt file; or a (named) list of such vectors. If not, use \code{\link{gprofiler2::gconvert}} to get required IDs.
family name. Example: human - 'hsapiens', mouse - 'mmusculus'.}

\item{ordered_query}{in case input gene lists are ranked this option may be
used to get GSEA style p-values.}

\item{multi_query}{in case of multiple gene lists, returns comparison table of these lists.
If enabled, the result data frame has columns named 'p_values', 'query_sizes', 'intersection_sizes' with vectors showing values in the order of input queries.}

\item{significant}{whether all or only statistically significant results should
be returned.}

\item{exclude_iea}{exclude GO electronic annotations (IEA).}

\item{measure_underrepresentation}{measure underrepresentation.}

\item{evcodes}{include evidence codes to the results. Note
that this can decrease performance and make the query slower.
In addition, a column 'intersection' is created that contains the gene id-s that intersect between the query and term.}

\item{user_threshold}{custom p-value threshold, results with a larger p-value are
excluded.}

\item{correction_method}{the algorithm used for multiple testing correction, one of "gSCS" (synonyms: "analytical", "g_SCS"), "fdr" (synonyms: "false_discovery_rate"), "bonferroni".}

\item{domain_scope}{how to define statistical domain, one of "annotated", "known".}

\item{custom_bg}{vector of gene names to use as a statistical background. If given, the domain_scope is set to 'custom'.}

\item{numeric_ns}{namespace to use for fully numeric IDs.}

\item{custom_gmt}{A path to the gmt file.}
}
\value{
A named list where 'result' contains data.frame with the enrichment analysis results and 'meta' contains metadata needed for Manhattan plot. If the input
 consisted of several lists the corresponding list is indicated with a variable
 'query'.
 When requesting a 'multi_query', either TRUE or FALSE, the columns of the resulting data frame differ.
 If 'evcodes' is set, the return value includes columns 'evidence_codes' and 'intersection'.
 The latter conveys info about the intersecting genes between the corresponding query and term.
}
\description{
Interface to the g:Profiler tool g:GOSt for functional enrichment of a list of gene IDs based on a given GMT file.
In case the input 'query' is a list, results for multiple queries will be returned in the same data frame with column 'query' indicating the corresponding query name.
If 'multi_query' is selected, the result is a data frame for comparing multiple input lists,
just as in the web tool.
}
\examples{
path_to_gmt <- file.path(path.package("gprofiler2.addon"),
                         "inst", "extdata", "h.all.v6.2.entrez.gmt")
query <- c("26118", "5837", "6781", "23036", "694", "123", "1466", "7436",
           "23210", "2131", "2152", "5165", "55139", "7360", "229", "8614",
           "54206", "2027", "10957", "3162", "5228", "26330", "9435", "55076")
gost_custom_gmt_res <- gost_custom_gmt(query, custom_gmt = path_to_gmt)

}
\author{
Gwenaëlle Lemoine <lemoine.gwenaelle@gmail.com>
}
