% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gost_custom_gmt_res_ref.R
\docType{data}
\name{gost_custom_gmt_res_ref}
\alias{gost_custom_gmt_res_ref}
\title{Reference gost_custom_gmt function result}
\format{An object of class \code{list} of length 2.}
\usage{
gost_custom_gmt_res_ref
}
\description{
Used in unit test to check if a same query return different results
}
\details{
Runned on 2019-06-26T22:48:25 on example into \code{\link{gost_custom_gmt}}
}
\keyword{datasets}
